zoom accuracy cleanup

+++ gifer/AnimationHelper.cs
@@ -2,6 +2,6 @@
 
 namespace gifer {
     public class AnimationHelper {
-        public static double GetEnlargementValue(double ratio) => Math.Abs(ratio > 0 ? ratio : 1 / ratio);
+        public static float GetEnlargementValue(float ratio) => Math.Abs(ratio > 0 ? ratio : 1 / ratio);
     }
 }
